New class GraphPlotting
All functions for the plotting were removed from Measurement_Properties and added to the new class GraphPlotting.
Added Errormessages if no SerialPort is selected.
