+++ Biodex Client/Biodex Client/Biodex Client.csproj
@@ -207,6 +207,7 @@
     <Compile Include="Biodex_Client.Designer.cs">
       <DependentUpon>Biodex_Client.cs</DependentUpon>
     </Compile>
+    <Compile Include="GraphPlotting.cs" />
     <Compile Include="Data.cs" />
     <Compile Include="DataAccessObject.cs" />
     <Compile Include="DB Classes\BiodexReport.cs" />
