+++ Biodex Client/Biodex Client/Data.cs
@@ -22,11 +22,11 @@ namespace Biodex_Client
         const double ANGLE_CAL_SUBTRAHEND = -99.3;
 
         //Measurment Properties + Calibrated Measurement data
-        public double[] Time { get; set; }
-        public double[] Torque { get; set; }
-        public double[] Velocity { get; set; }
-        public double[] Angle { get; set; }
-        public Data GraphDataCalibrated { get; set; }
+        //public double[] Time { get; set; }
+        //public double[] Torque { get; set; }
+        //public double[] Velocity { get; set; }
+        //public double[] Angle { get; set; }
+        //public Data GraphDataCalibrated { get; set; }
 
         //Measurment Properties as Lists
         public List<double> aTimeList { get; set; }
@@ -42,14 +42,14 @@ namespace Biodex_Client
             aAngleList = new List<double>();
         }
 
-        public Data(int[][] data) : this()
-        {
-            Time = createTimeProperty(data, Time);
-            Torque = getDatainClassProperty(data, Torque, 0);
-            Velocity = getDatainClassProperty(data, Velocity, 1);
-            Angle = getDatainClassProperty(data, Angle, 2);
-            GraphDataCalibrated = calibrateData(data, Time, Torque, Velocity, Angle);
-        }
+        //public Data(int[][] data) : this()
+        //{
+        //    Time = createTimeProperty(data, Time);
+        //    Torque = getDatainClassProperty(data, Torque, 0);
+        //    Velocity = getDatainClassProperty(data, Velocity, 1);
+        //    Angle = getDatainClassProperty(data, Angle, 2);
+        //    GraphDataCalibrated = calibrateData(data, Time, Torque, Velocity, Angle);
+        //}
 
         /*
         * adds single data values to the lists
@@ -67,7 +67,7 @@ namespace Biodex_Client
         {
 			try
 			{
-                Biodex_Client.FormMeasurementProperties.refreshCharts();
+                Biodex_Client.GraphPlotting.refreshCharts();
 
                 aTorqueList = Array.ConvertAll(Torque.Split(';'), double.Parse).ToList();
                 aVelocityList = Array.ConvertAll(Velocity.Split(';'), double.Parse).ToList();
@@ -78,9 +78,9 @@ namespace Biodex_Client
                 {
                     aTimeList.Add(aTimeList.Count + 1);
 
-                    Biodex_Client.FormMeasurementProperties.ChartValuesTorqueValues.Add(new formMeasurementProperties.ValuePoint(aTimeList[i], aTorqueList[i]));
-                    Biodex_Client.FormMeasurementProperties.ChartValuesVelocityValues.Add(new formMeasurementProperties.ValuePoint(aTimeList[i], aVelocityList[i]));
-                    Biodex_Client.FormMeasurementProperties.ChartValuesAngleValues.Add(new formMeasurementProperties.ValuePoint(aTimeList[i], aAngleList[i]));
+                    Biodex_Client.GraphPlotting.ChartValuesTorqueValues.Add(new ValuePoint(aTimeList[i], aTorqueList[i]));
+                    Biodex_Client.GraphPlotting.ChartValuesVelocityValues.Add(new ValuePoint(aTimeList[i], aVelocityList[i]));
+                    Biodex_Client.GraphPlotting.ChartValuesAngleValues.Add(new ValuePoint(aTimeList[i], aAngleList[i]));
                 }
 
                 MessageBox.Show("Loaded Former Records From Database Successfully and PLOTTED THEM. Now You Can Also Create a CSV-File", "Loading Done", MessageBoxButtons.OK, MessageBoxIcon.Information);
@@ -106,52 +106,52 @@ namespace Biodex_Client
         /*
          * generates the calibrated data object 
          */
-        Data calibrateData(int[][] data, double[] time, double[] torque, double[] velocity, double[] angle)
-        {
-            Data graphData = new Data();
-            graphData.Time = time;
-            graphData.Torque = torque;
-            graphData.Velocity = velocity;
-            graphData.Angle = angle;
+        //Data calibrateData(int[][] data, double[] time, double[] torque, double[] velocity, double[] angle)
+        //{
+        //    Data graphData = new Data();
+        //    graphData.Time = time;
+        //    graphData.Torque = torque;
+        //    graphData.Velocity = velocity;
+        //    graphData.Angle = angle;
 
-            for (int i = 0; i < data[0].Length; i++)
-            {
-                graphData.Torque[i] = graphData.Torque[i] * TORQUE_CAL_FACTOR + TORQUE_CAL_SUBTRAHEND;
-                graphData.Velocity[i] = graphData.Velocity[i] * VELOCITY_CAL_FACTOR + VELOCITY_CAL_SUBTRAHEND;
-                graphData.Angle[i] = graphData.Angle[i] * ANGLE_CAL_FACTOR + ANGLE_CAL_SUBTRAHEND;
-            }
-            return graphData;
-        }
+        //    for (int i = 0; i < data[0].Length; i++)
+        //    {
+        //        graphData.Torque[i] = graphData.Torque[i] * TORQUE_CAL_FACTOR + TORQUE_CAL_SUBTRAHEND;
+        //        graphData.Velocity[i] = graphData.Velocity[i] * VELOCITY_CAL_FACTOR + VELOCITY_CAL_SUBTRAHEND;
+        //        graphData.Angle[i] = graphData.Angle[i] * ANGLE_CAL_FACTOR + ANGLE_CAL_SUBTRAHEND;
+        //    }
+        //    return graphData;
+        //}
 
         /*
          * arranges data from data array into data object properties
          * will be removed as loadbutton and serialport are available
          */
-        double[] getDatainClassProperty(int[][] data, double[] classData, int index)
-        {
-            double[] temp = new double[data[0].Length];
-            for (int i = 0; i < data[0].Length; i++)
-            {
-                temp[i] = Convert.ToDouble(data[index][i]);
-            }
-            classData = temp;
-            return classData;
-        }
+        //double[] getDatainClassProperty(int[][] data, double[] classData, int index)
+        //{
+        //    double[] temp = new double[data[0].Length];
+        //    for (int i = 0; i < data[0].Length; i++)
+        //    {
+        //        temp[i] = Convert.ToDouble(data[index][i]);
+        //    }
+        //    classData = temp;
+        //    return classData;
+        //}
 
         /*
          * generates the data object time/frames property
          * will be changed when loadbutton and serialport are implemented
          */
-        double[] createTimeProperty(int[][] data, double[] classData)
-        {
-            double[] temp = new double[data[0].Length];
-            for (int i = 0; i < data[0].Length; i++)
-            {
+    //    double[] createTimeProperty(int[][] data, double[] classData)
+    //    {
+    //        double[] temp = new double[data[0].Length];
+    //        for (int i = 0; i < data[0].Length; i++)
+    //        {
 
-                temp[i] = i;
-            }
-            classData = temp;
-            return classData;
-        }
+    //            temp[i] = i;
+    //        }
+    //        classData = temp;
+    //        return classData;
+    //    }
     }
 }
